feat: Add CreateUserTx method to SQLC store
This commit adds the CreateUserTx method to the SQLC store interface in the db package. The CreateUserTx method allows creating a new user transaction with the provided parameters. It takes a gin.Context, a *pgx.Conn, and CreateUserTxParams as input and returns a *User and an error. This addition enhances the functionality of the SQLC store and enables the creation of user transactions in the application.