feat: Add migration files for adding user role column
The migration files `000004_add_user_role.up.sql` and `000004_add_user_role.down.sql` are added to the `db/migration` directory. These files contain the necessary SQL statements to add and remove the `role` column in the `users` table. This addition enhances the database schema and allows for the management of user roles in the application.