feat: Add role column to User model and SQL queries
This commit adds the `role` column to the `User` model in the `models.go` file. It also updates the SQL queries in the `user.sql` and `user.sql.go` files to include the `role` column in the `INSERT` and `UPDATE` statements. This modification allows for the management of user roles in the application and enhances the database schema.