chore: Update Terraform and Go versions in CI workflow
This commit updates the `ci.yaml` workflow in the `.github/workflows` directory to use the latest versions of Terraform and Go. The `TF_VERSION` environment variable is removed, and the `go-version` input for the `setup-go` action is updated to use the `$GO_VERSION` variable. This modification ensures that the CI workflow is using the most up-to-date dependencies and improves compatibility with the project.