chore: Update RDS configuration in infra
The `rds.tf` file in the `infra` directory has been modified to update the configuration of the RDS instance. The `allocated_storage` parameter has been changed to 10, the `db_name` has been set to "simple-bank", and the `engine_version` has been updated to "16.0". Additionally, the `skip_final_snapshot` parameter has been set to true. This change ensures that the RDS instance is properly configured for the application's needs.